Maintenance: Rename GnosisNFTTransfer to TransferNFTSection
