+++ src/modules/dashboard/components/Dialogs/ControlSafeDialog/TransactionTypesSection/index.ts
@@ -1,4 +1,4 @@
 export { default as TransferFundsSection } from './TransferFundsSection';
 export { default as RawTransactionSection } from './RawTransactionSection';
 export { default as ContractInteractionSection } from './ContractInteractionSection';
-export { default as GnosisNFTTransfer } from './GnosisNFTTransfer';
+export { default as TransferNFTSection } from './TransferNFTSection';
